refactor(pdfitem): make to own component
